backend/pyproject.toml
[tool.poetry]
name = "task-queue"
version = "0.1.0"
description = ""
authors = ["Dawid <rygula.dawid@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
ipython = "^8.22.2"
isort = {extras = ["black"], version = "^5.13.2"}
psycopg2 = "^2.9.9"
redis = "^5.0.3"
celery = "^5.3.6"
djangorestframework = "^3.15.0"
djangorestframework-camel-case = "^1.4.2"
drf-spectacular = "^0.27.1"
factory-boy = "^3.3.0"
sqids = "^0.4.1"

[tool.poetry.group.dev.dependencies]
django = "4.2"
black = "^24.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
